fix: loosen LinkAttribute length checks
Loosen the LinkAttribute length checks, to parse links that store
information other than MAC addresses in the Address or Broadcast fields.

This change, for example, now allows IPIP, GRE over IP, 6to4, and IP
over Infiniband.

The fields continue to be net.HardwareAddr, which may result in slightly
unusual representations when printed as a string. This type can be cast
to other byte slice types by users, if necessary.

Fix: #7