feat: add OperationalState and LinkInfo attributes
Add support for parsing two additional LinkAttributes necessary for
creating and managing interfaces: OperationState and LinkInfo.

OperationalState allows for the UP/DOWN state of the interface to be
determined and controlled. This addition allows for interfaces to be
created and then brought up programatically.

LinkInfo contains driver specific configuration, such as those for
bridges or tunnels. This additional configuration is stored as nested
Netlink messages on the Data and SlaveData members.