feat(route): list with kernel filtering
A new "ListMatch" function has been added to the RouteService to allow
for kernel-based filtering of the routing tables. The netlink flags
required for getting a route for an IP (eg, rtnl.RouteGetAll) is
different than those required while filtering the table.

This changeset removes the superfluous "netlink.DumpFiltered" flag from
the Get function of RouteService, as this flag is for kernel responses.

Signed-off-by: Terin Stock <terinjokes@gmail.com>