rtnetlink: fix Link ListByKind response
the linux kernel handles a RTM_GETLINK request even in case of incorrect
kind filter. if it finds requested kind it responds with
a list of this kind of links and marks every message in a response
by NLM_F_DUMP_FILTERED flag. Otherwise it sends all available links
w/o NLF_F_DUMP_FILTERED. So, looks much better to return to a callee
empty link list if there are no links of requested kind.

Signed-off-by: Andrey Grafin <conquistador@yandex-team.ru>